+++ src/Cluster.h
@@ -20,7 +20,7 @@ public:
 	void agregarElemento(Punto* elemento);
 	Punto getCentroide();
 	bool esSimilar(Cluster cluster);
-	double gradoCohesion();
+	float gradoCohesion();
 	vector<Punto*> getPuntos();
 	void vaciarPuntos();
